Allow git dependencies.
One problem with running node2nix within the build is that it fails if
there are git dependencies due to the network sandbox. This is resolved
by following the pattern of running node2nix in a `generate.sh` script
prior to building. The only difference in the generated expressions
is the relative path which `src` points to, so this must be overridden
at build time.

Bootstrap from the master branch is added as a demonstration.